Move DB setup to shared location so tests can also use it
