+++ cmd/bank/main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"context"
-	"database/sql"
 	"os"
 	"time"
 
@@ -10,9 +9,9 @@ import (
 
 	"github.com/dogmatiq/dogma"
 	"github.com/dogmatiq/example"
+	"github.com/dogmatiq/example/internal/database"
 	"github.com/dogmatiq/example/messages"
 	"github.com/dogmatiq/example/messages/commands"
-	"github.com/dogmatiq/projectionkit/sql/sqlite"
 	"github.com/dogmatiq/testkit/engine"
 )
 
@@ -21,7 +20,7 @@ func businessDayFromTime(t time.Time) string {
 }
 
 func main() {
-	app, err := example.NewApp(newDB())
+	app, err := example.NewApp(database.New())
 	if err != nil {
 		panic(err)
 	}
@@ -91,35 +90,4 @@ func main() {
 	if err := app.GenerateAccountCSV(os.Stdout); err != nil {
 		panic(err)
 	}
-}
-
-func newDB() *sql.DB {
-	ctx := context.Background()
-
-	db, err := sql.Open("sqlite3", "file::memory:?cache=shared")
-	if err != nil {
-		panic(err)
-	}
-	defer db.Close()
-
-	db.SetMaxIdleConns(1)
-	db.SetConnMaxLifetime(-1)
-
-	if err := sqlite.CreateSchema(ctx, db); err != nil {
-		panic(err)
-	}
-
-	if _, err := db.ExecContext(
-		ctx,
-		`CREATE TABLE customer (
-			id   TEXT NOT NULL,
-			name TEXT NOT NULL,
-
-			PRIMARY KEY (id)
-		)`,
-	); err != nil {
-		panic(err)
-	}
-
-	return db
 }
