+++ internal/testrunner/testrunner.go
@@ -2,8 +2,18 @@ package testrunner
 
 import (
 	"github.com/dogmatiq/example"
+	"github.com/dogmatiq/example/internal/database"
 	"github.com/dogmatiq/testkit"
 )
 
 // Runner is a test runner for the example app.
-var Runner = testkit.New(&example.App{})
+var Runner *testkit.Runner
+
+func init() {
+	app, err := example.NewApp(database.New())
+	if err != nil {
+		panic(err)
+	}
+
+	Runner = testkit.New(app)
+}
